Realocação da classe Terra e configurações da classe Ground
